+++ back/src/database/models/meal_food.js
@@ -37,17 +37,17 @@ const MealFood = sequelize.define('Meal_Food', {
     defaultValue: 0
   },
   Prot: {
-    type: DataTypes.FLOAT(4,1),
+    type: DataTypes.FLOAT,
     allowNull: false,
     defaultValue: 0
   },
   Carb: {
-    type: DataTypes.FLOAT(4,1),
+    type: DataTypes.FLOAT,
     allowNull: false,
     defaultValue: 0
   },
   Fat: {
-    type: DataTypes.FLOAT(4,1),
+    type: DataTypes.FLOAT,
     allowNull: false,
     defaultValue: 0
   },
